src/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
flask==3.0.3
gunicorn==22.0.0
idna==3.10
numpy==2.2.5
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
scikit-learn==1.5.0
six==1.17.0
spotipy==2.25.1
tzdata==2025.2
urllib3==2.4.0
shiny==0.8.1